# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: Cabales API
  version: 1.2.0
  license: { name: ISC, identifier: ISC }
  description: >-
    API de autenticación, grupos, eventos, gastos exactos, liquidaciones, fondos, presupuestos, documentos, OCR, Cabudas, estadísticas, avisos, logros y privacidad. Importes en centavos.
    Todas las respuestas /api/v1 usan Cache-Control private, no-store; health y ready usan no-store.
servers:
  - url: /
...
paths:
...
  /api/v1/documents:
    get:
      tags: [Documents]
      summary: 'Documentos accesibles (propios, concedidos o del grupo)'
      operationId: listDocuments
      parameters:
        - name: groupId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: eventId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: expenseId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: settlementId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: cursor
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 30
      security:
        - sessionCookie: []
      responses:
        '200':
          description: 'Operación correcta'
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    const: true
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Document' }
                  meta:
                    type: object
                    properties:
                      nextCursor:
                        type: [string, 'null']
                        format: uuid
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
    post:
      tags: [Documents]
      summary: 'Sube un archivo binario (PDF, JPEG, PNG o WebP)'
      operationId: uploadDocument
      description: 'El Content-Type debe ser application/pdf, image/jpeg, image/png o image/webp y coincidir con la firma binaria (415/422 CONTENT_MISMATCH). Tamaño máximo MAX_UPLOAD_BYTES. La clave de almacenamiento nunca se expone.'
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 255
        - name: groupId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: eventId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: expenseId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: settlementId
          in: query
          required: false
          schema:
            type: string
            format: uuid
      security:
        - sessionCookie: []
          csrfCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              contentMediaType: application/octet-stream
      responses:
        '201':
          description: 'Operación correcta'
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    const: true
                  data: { $ref: '#/components/schemas/Document' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /api/v1/documents/{documentId}:
    parameters:
      - $ref: '#/components/parameters/DocumentId'
    get:
      tags: [Documents]
      summary: 'Metadatos de un documento'
      operationId: getDocument
      security:
        - sessionCookie: []
      responses:
        '200':
          description: 'Operación correcta'
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    const: true
                  data: { $ref: '#/components/schemas/Document' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
    patch:
      tags: [Documents]
      summary: 'Renombra o cambia asociaciones (acceso EDIT)'
      operationId: updateDocument
      security:
        - sessionCookie: []
          csrfCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DocumentPatch' }
      responses:
        '200':
          description: 'Operación correcta'
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    const: true
                  data: { $ref: '#/components/schemas/Document' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
    delete:
      tags: [Documents]
      summary: 'Elimina el documento y su archivo (acceso MANAGE)'
      operationId: deleteDocument
      security:
        - sessionCookie: []
          csrfCookie: []
          csrfHeader: []
      responses:
        '200':
          description: 'Operación correcta'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MutationEnvelope' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /api/v1/documents/{documentId}/download-url:
    parameters:
      - $ref: '#/components/parameters/DocumentId'
    post:
      tags: [Documents]
      summary: 'Emite una URL firmada de corta duración'
      operationId: createDocumentDownloadUrl
      security:
        - sessionCookie: []
          csrfCookie: []
          csrfHeader: []
      responses:
        '200':
          description: 'Operación correcta'
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    const: true
                  data:
                    type: object
                    required: [url, expiresAt]
                    properties:
                      url:
                        type: string
                        format: uri
                      expiresAt:
                        type: string
                        format: date-time
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /api/v1/documents/{documentId}/grants:
    parameters:
      - $ref: '#/components/parameters/DocumentId'
    get:
      tags: [Documents]
      summary: 'Permisos explícitos del documento (acceso MANAGE)'
      operationId: listDocumentGrants
      security:
        - sessionCookie: []
      responses:
        '200':
          description: 'Operación correcta'
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    const: true
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/DocumentGrant' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /api/v1/documents/{documentId}/grants/{userId}:
    parameters:
      - $ref: '#/components/parameters/DocumentId'
      - $ref: '#/components/parameters/UserId'
    put:
      tags: [Documents]
      summary: 'Concede o cambia acceso a un miembro del grupo'
      operationId: putDocumentGrant
      security:
        - sessionCookie: []
          csrfCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [access]
              properties:
                access:
                  enum: [VIEW, EDIT, MANAGE]
      responses:
        '200':
          description: 'Operación correcta'
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    const: true
                  data: { $ref: '#/components/schemas/DocumentGrant' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
    delete:
      tags: [Documents]
      summary: 'Retira un permiso explícito'
      operationId: deleteDocumentGrant
      security:
        - sessionCookie: []
          csrfCookie: []
          csrfHeader: []
      responses:
        '200':
          description: 'Operación correcta'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MutationEnvelope' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /api/v1/documents/{documentId}/access-logs:
    parameters:
      - $ref: '#/components/parameters/DocumentId'
    get:
      tags: [Documents]
      summary: 'Bitácora de accesos (acceso MANAGE)'
      operationId: listDocumentAccessLogs
      security:
        - sessionCookie: []
      responses:
        '200':
          description: 'Operación correcta'
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    const: true
                  data:
                    type: array
                    items:
                      type: object
                      required: [id, action, createdAt, user]
                      properties:
                        id:
                          type: string
                          format: uuid
                        action:
                          type: string
                        createdAt:
                          type: string
                          format: date-time
                        user:
                          type: [object, 'null']
                          properties:
                            id:
                              type: string
                              format: uuid
                            displayName:
                              type: string
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: cabales_session
      description: Nombre predeterminado configurable mediante COOKIE_NAME.
    csrfCookie:
      type: apiKey
      in: cookie
      name: cabales_session_csrf
      description: Cookie de doble envío cuyo prefijo sigue COOKIE_NAME; /auth/me valida su hash antes de devolverla.
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Debe coincidir con csrfCookie y con el hash ligado a la sesión.
  parameters:
...
    DocumentId:
      name: documentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    BadRequest:
      {
        description: Entrada inválida,
        content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } },
      }
    Unauthorized:
      {
        description: Sesión ausente o inválida,
        content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } },
      }
    Forbidden:
      {
        description: CSRF o permiso inválido,
        content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } },
      }
    NotFound:
      {
        description: Recurso no visible,
        content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } },
      }
...
    UnprocessableEntity:
      {
        description: Invariante de negocio incumplida,
        content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } },
      }
    TooManyRequests:
      {
        description: Límite de solicitudes excedido,
        content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } },
      }
    PayloadTooLarge:
      description: 'Archivo o cuerpo demasiado grande'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    UnsupportedMediaType:
      description: 'Tipo de contenido no permitido'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    ServiceUnavailable:
      {
        description: PostgreSQL no disponible,
        content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } },
      }
  schemas:
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [success, error]
      properties:
        success: { const: false }
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
            details: {}
...
    MutationEnvelope:
      type: object
      required: [success, data]
      properties:
        {
          success: { const: true },
          data: { type: object, additionalProperties: { type: boolean } },
        }
...
    Document:
      type: object
      required:
        [
          id,
          name,
          mimeType,
          sizeBytes,
          groupId,
          eventId,
          expenseId,
          settlementId,
          createdAt,
          updatedAt,
          owner,
          access,
        ]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        mimeType:
          enum: [application/pdf, image/jpeg, image/png, image/webp]
        sizeBytes:
          type: [integer, 'null']
        groupId:
          type: [string, 'null']
          format: uuid
        eventId:
          type: [string, 'null']
          format: uuid
        expenseId:
          type: [string, 'null']
          format: uuid
        settlementId:
          type: [string, 'null']
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        owner:
          type: object
          required: [id, displayName]
          properties:
            id:
              type: string
              format: uuid
            displayName:
              type: string
        access:
          enum: [VIEW, EDIT, MANAGE]
    DocumentPatch:
      type: object
      additionalProperties: false
      required: []
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        eventId:
          type: [string, 'null']
          format: uuid
        expenseId:
          type: [string, 'null']
          format: uuid
        settlementId:
          type: [string, 'null']
          format: uuid
      minProperties: 1
    DocumentGrant:
      type: object
      required: [userId, displayName, access, createdAt]
      properties:
        userId:
          type: string
          format: uuid
        displayName:
          type: string
        access:
          enum: [VIEW, EDIT, MANAGE]
        createdAt:
          type: string
          format: date-time
...